Refactor draggable component to DraggableComponent
